[ads] Cleanup brave://rewards-internals diagnostics
