+++ components/brave_ads/core/internal/BUILD.gn
@@ -623,14 +623,14 @@ static_library("internal") {
     "diagnostics/entries/last_unidle_time_diagnostic_util.h",
     "diagnostics/entries/locale_diagnostic_entry.cc",
     "diagnostics/entries/locale_diagnostic_entry.h",
-    "diagnostics/entries/opted_in_to_brave_news_ads_diagnostic_entry.cc",
-    "diagnostics/entries/opted_in_to_brave_news_ads_diagnostic_entry.h",
-    "diagnostics/entries/opted_in_to_new_tab_page_ads_diagnostic_entry.cc",
-    "diagnostics/entries/opted_in_to_new_tab_page_ads_diagnostic_entry.h",
-    "diagnostics/entries/opted_in_to_notification_ads_diagnostic_entry.cc",
-    "diagnostics/entries/opted_in_to_notification_ads_diagnostic_entry.h",
-    "diagnostics/entries/opted_in_to_search_result_ads_diagnostic_entry.cc",
-    "diagnostics/entries/opted_in_to_search_result_ads_diagnostic_entry.h",
+    "diagnostics/entries/opted_into_brave_news_ads_diagnostic_entry.cc",
+    "diagnostics/entries/opted_into_brave_news_ads_diagnostic_entry.h",
+    "diagnostics/entries/opted_into_new_tab_page_ads_diagnostic_entry.cc",
+    "diagnostics/entries/opted_into_new_tab_page_ads_diagnostic_entry.h",
+    "diagnostics/entries/opted_into_notification_ads_diagnostic_entry.cc",
+    "diagnostics/entries/opted_into_notification_ads_diagnostic_entry.h",
+    "diagnostics/entries/opted_into_search_result_ads_diagnostic_entry.cc",
+    "diagnostics/entries/opted_into_search_result_ads_diagnostic_entry.h",
     "flags/debug/debug_command_line_switch_parser_util.cc",
     "flags/debug/debug_command_line_switch_parser_util.h",
     "flags/debug/debug_flag_util.cc",
