+++ components/brave_ads/core/internal/diagnostics/entries/opted_into_brave_news_ads_diagnostic_entry.cc
@@ -3,7 +3,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this file,
  * You can obtain one at https://mozilla.org/MPL/2.0/. */
 
-#include "brave/components/brave_ads/core/internal/diagnostics/entries/opted_in_to_new_tab_page_ads_diagnostic_entry.h"
+#include "brave/components/brave_ads/core/internal/diagnostics/entries/opted_into_brave_news_ads_diagnostic_entry.h"
 
 #include "brave/components/brave_ads/core/internal/common/strings/string_conversions_util.h"
 #include "brave/components/brave_ads/core/internal/settings/settings.h"
@@ -11,19 +11,19 @@
 namespace brave_ads {
 
 namespace {
-constexpr char kName[] = "Opted-in to new tab page ads";
+constexpr char kName[] = "Opted into Brave News ads";
 }  // namespace
 
-DiagnosticEntryType OptedInToNewTabPageAdsDiagnosticEntry::GetType() const {
-  return DiagnosticEntryType::kOptedInToNewTabPageAds;
+DiagnosticEntryType OptedInToBraveNewsAdsDiagnosticEntry::GetType() const {
+  return DiagnosticEntryType::kOptedInToBraveNewsAds;
 }
 
-std::string OptedInToNewTabPageAdsDiagnosticEntry::GetName() const {
+std::string OptedInToBraveNewsAdsDiagnosticEntry::GetName() const {
   return kName;
 }
 
-std::string OptedInToNewTabPageAdsDiagnosticEntry::GetValue() const {
-  return BoolToString(UserHasOptedInToNewTabPageAds());
+std::string OptedInToBraveNewsAdsDiagnosticEntry::GetValue() const {
+  return BoolToString(UserHasOptedInToBraveNewsAds());
 }
 
 }  // namespace brave_ads
