+++ components/brave_ads/core/internal/diagnostics/entries/opted_into_new_tab_page_ads_diagnostic_entry.cc
@@ -3,7 +3,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this file,
  * You can obtain one at https://mozilla.org/MPL/2.0/. */
 
-#include "brave/components/brave_ads/core/internal/diagnostics/entries/opted_in_to_notification_ads_diagnostic_entry.h"
+#include "brave/components/brave_ads/core/internal/diagnostics/entries/opted_into_new_tab_page_ads_diagnostic_entry.h"
 
 #include "brave/components/brave_ads/core/internal/common/strings/string_conversions_util.h"
 #include "brave/components/brave_ads/core/internal/settings/settings.h"
@@ -11,19 +11,19 @@
 namespace brave_ads {
 
 namespace {
-constexpr char kName[] = "Opted-in to notification ads";
+constexpr char kName[] = "Opted into new tab page ads";
 }  // namespace
 
-DiagnosticEntryType OptedInToNotificationAdsDiagnosticEntry::GetType() const {
-  return DiagnosticEntryType::kOptedInToNotificationAds;
+DiagnosticEntryType OptedInToNewTabPageAdsDiagnosticEntry::GetType() const {
+  return DiagnosticEntryType::kOptedInToNewTabPageAds;
 }
 
-std::string OptedInToNotificationAdsDiagnosticEntry::GetName() const {
+std::string OptedInToNewTabPageAdsDiagnosticEntry::GetName() const {
   return kName;
 }
 
-std::string OptedInToNotificationAdsDiagnosticEntry::GetValue() const {
-  return BoolToString(UserHasOptedInToNotificationAds());
+std::string OptedInToNewTabPageAdsDiagnosticEntry::GetValue() const {
+  return BoolToString(UserHasOptedInToNewTabPageAds());
 }
 
 }  // namespace brave_ads
