+++ components/brave_ads/core/internal/diagnostics/entries/opted_into_notification_ads_diagnostic_entry.cc
@@ -3,7 +3,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this file,
  * You can obtain one at https://mozilla.org/MPL/2.0/. */
 
-#include "brave/components/brave_ads/core/internal/diagnostics/entries/opted_in_to_brave_news_ads_diagnostic_entry.h"
+#include "brave/components/brave_ads/core/internal/diagnostics/entries/opted_into_notification_ads_diagnostic_entry.h"
 
 #include "brave/components/brave_ads/core/internal/common/strings/string_conversions_util.h"
 #include "brave/components/brave_ads/core/internal/settings/settings.h"
@@ -11,19 +11,19 @@
 namespace brave_ads {
 
 namespace {
-constexpr char kName[] = "Opted-in to Brave News ads";
+constexpr char kName[] = "Opted into notification ads";
 }  // namespace
 
-DiagnosticEntryType OptedInToBraveNewsAdsDiagnosticEntry::GetType() const {
-  return DiagnosticEntryType::kOptedInToBraveNewsAds;
+DiagnosticEntryType OptedInToNotificationAdsDiagnosticEntry::GetType() const {
+  return DiagnosticEntryType::kOptedInToNotificationAds;
 }
 
-std::string OptedInToBraveNewsAdsDiagnosticEntry::GetName() const {
+std::string OptedInToNotificationAdsDiagnosticEntry::GetName() const {
   return kName;
 }
 
-std::string OptedInToBraveNewsAdsDiagnosticEntry::GetValue() const {
-  return BoolToString(UserHasOptedInToBraveNewsAds());
+std::string OptedInToNotificationAdsDiagnosticEntry::GetValue() const {
+  return BoolToString(UserHasOptedInToNotificationAds());
 }
 
 }  // namespace brave_ads
