+++ components/brave_ads/core/internal/diagnostics/entries/opted_into_notification_ads_diagnostic_entry.h
@@ -3,8 +3,8 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this file,
  * You can obtain one at https://mozilla.org/MPL/2.0/. */
 
-#ifndef BRAVE_COMPONENTS_BRAVE_ADS_CORE_INTERNAL_DIAGNOSTICS_ENTRIES_OPTED_IN_TO_NEW_TAB_PAGE_ADS_DIAGNOSTIC_ENTRY_H_
-#define BRAVE_COMPONENTS_BRAVE_ADS_CORE_INTERNAL_DIAGNOSTICS_ENTRIES_OPTED_IN_TO_NEW_TAB_PAGE_ADS_DIAGNOSTIC_ENTRY_H_
+#ifndef BRAVE_COMPONENTS_BRAVE_ADS_CORE_INTERNAL_DIAGNOSTICS_ENTRIES_OPTED_INTO_NOTIFICATION_ADS_DIAGNOSTIC_ENTRY_H_
+#define BRAVE_COMPONENTS_BRAVE_ADS_CORE_INTERNAL_DIAGNOSTICS_ENTRIES_OPTED_INTO_NOTIFICATION_ADS_DIAGNOSTIC_ENTRY_H_
 
 #include <string>
 
@@ -12,7 +12,7 @@
 
 namespace brave_ads {
 
-class OptedInToNewTabPageAdsDiagnosticEntry final
+class OptedInToNotificationAdsDiagnosticEntry final
     : public DiagnosticEntryInterface {
  public:
   // DiagnosticEntryInterface:
@@ -23,4 +23,4 @@ class OptedInToNewTabPageAdsDiagnosticEntry final
 
 }  // namespace brave_ads
 
-#endif  // BRAVE_COMPONENTS_BRAVE_ADS_CORE_INTERNAL_DIAGNOSTICS_ENTRIES_OPTED_IN_TO_NEW_TAB_PAGE_ADS_DIAGNOSTIC_ENTRY_H_
+#endif  // BRAVE_COMPONENTS_BRAVE_ADS_CORE_INTERNAL_DIAGNOSTICS_ENTRIES_OPTED_INTO_NOTIFICATION_ADS_DIAGNOSTIC_ENTRY_H_
