+++ components/brave_ads/core/internal/diagnostics/entries/opted_into_search_result_ads_diagnostic_entry.cc
@@ -3,7 +3,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this file,
  * You can obtain one at https://mozilla.org/MPL/2.0/. */
 
-#include "brave/components/brave_ads/core/internal/diagnostics/entries/opted_in_to_search_result_ads_diagnostic_entry.h"
+#include "brave/components/brave_ads/core/internal/diagnostics/entries/opted_into_search_result_ads_diagnostic_entry.h"
 
 #include "brave/components/brave_ads/core/internal/common/strings/string_conversions_util.h"
 #include "brave/components/brave_ads/core/internal/settings/settings.h"
@@ -11,7 +11,7 @@
 namespace brave_ads {
 
 namespace {
-constexpr char kName[] = "Opted-in to search result ads";
+constexpr char kName[] = "Opted into search result ads";
 }  // namespace
 
 DiagnosticEntryType OptedInToSearchResultAdsDiagnosticEntry::GetType() const {
