+++ components/brave_ads/core/internal/diagnostics/entries/opted_into_search_result_ads_diagnostic_entry.h
@@ -3,8 +3,8 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this file,
  * You can obtain one at https://mozilla.org/MPL/2.0/. */
 
-#ifndef BRAVE_COMPONENTS_BRAVE_ADS_CORE_INTERNAL_DIAGNOSTICS_ENTRIES_OPTED_IN_TO_SEARCH_RESULT_ADS_DIAGNOSTIC_ENTRY_H_
-#define BRAVE_COMPONENTS_BRAVE_ADS_CORE_INTERNAL_DIAGNOSTICS_ENTRIES_OPTED_IN_TO_SEARCH_RESULT_ADS_DIAGNOSTIC_ENTRY_H_
+#ifndef BRAVE_COMPONENTS_BRAVE_ADS_CORE_INTERNAL_DIAGNOSTICS_ENTRIES_OPTED_INTO_SEARCH_RESULT_ADS_DIAGNOSTIC_ENTRY_H_
+#define BRAVE_COMPONENTS_BRAVE_ADS_CORE_INTERNAL_DIAGNOSTICS_ENTRIES_OPTED_INTO_SEARCH_RESULT_ADS_DIAGNOSTIC_ENTRY_H_
 
 #include <string>
 
@@ -23,4 +23,4 @@ class OptedInToSearchResultAdsDiagnosticEntry final
 
 }  // namespace brave_ads
 
-#endif  // BRAVE_COMPONENTS_BRAVE_ADS_CORE_INTERNAL_DIAGNOSTICS_ENTRIES_OPTED_IN_TO_SEARCH_RESULT_ADS_DIAGNOSTIC_ENTRY_H_
+#endif  // BRAVE_COMPONENTS_BRAVE_ADS_CORE_INTERNAL_DIAGNOSTICS_ENTRIES_OPTED_INTO_SEARCH_RESULT_ADS_DIAGNOSTIC_ENTRY_H_
