+++ components/brave_ads/core/test/BUILD.gn
@@ -374,10 +374,10 @@ source_set("brave_ads_unit_tests") {
     "//brave/components/brave_ads/core/internal/diagnostics/entries/device_id_diagnostic_entry_unittest.cc",
     "//brave/components/brave_ads/core/internal/diagnostics/entries/last_unidle_time_diagnostic_entry_unittest.cc",
     "//brave/components/brave_ads/core/internal/diagnostics/entries/locale_diagnostic_entry_unittest.cc",
-    "//brave/components/brave_ads/core/internal/diagnostics/entries/opted_in_to_brave_news_ads_diagnostic_entry_unittest.cc",
-    "//brave/components/brave_ads/core/internal/diagnostics/entries/opted_in_to_new_tab_page_ads_diagnostic_entry_unittest.cc",
-    "//brave/components/brave_ads/core/internal/diagnostics/entries/opted_in_to_notification_ads_diagnostic_entry_unittest.cc",
-    "//brave/components/brave_ads/core/internal/diagnostics/entries/opted_in_to_search_result_ads_diagnostic_entry_unittest.cc",
+    "//brave/components/brave_ads/core/internal/diagnostics/entries/opted_into_brave_news_ads_diagnostic_entry_unittest.cc",
+    "//brave/components/brave_ads/core/internal/diagnostics/entries/opted_into_new_tab_page_ads_diagnostic_entry_unittest.cc",
+    "//brave/components/brave_ads/core/internal/diagnostics/entries/opted_into_notification_ads_diagnostic_entry_unittest.cc",
+    "//brave/components/brave_ads/core/internal/diagnostics/entries/opted_into_search_result_ads_diagnostic_entry_unittest.cc",
     "//brave/components/brave_ads/core/internal/flags/debug/debug_command_line_switch_parser_util_unittest.cc",
     "//brave/components/brave_ads/core/internal/flags/did_override/did_override_command_line_switch_values_util_unittest.cc",
     "//brave/components/brave_ads/core/internal/flags/did_override/did_override_command_line_switches_util_unittest.cc",
